Remove Bundle from namespace
