+++ src/OpenConext/EngineBlockApiClient/Exception/InvalidArgumentException.php
@@ -18,8 +18,10 @@ declare(strict_types = 1);
  * limitations under the License.
  */
 
-namespace OpenConext\EngineBlockApiClientBundle\Exception;
+namespace OpenConext\EngineBlockApiClient\Exception;
 
-interface OpenConextEngineBlockApiClientException
+use InvalidArgumentException as CoreInvalidArgumentException;
+
+class InvalidArgumentException extends CoreInvalidArgumentException implements OpenConextEngineBlockApiClientException
 {
 }
