+++ src/OpenConext/EngineBlockApiClient/Exception/RuntimeException.php
@@ -18,8 +18,10 @@ declare(strict_types = 1);
  * limitations under the License.
  */
 
-namespace OpenConext\EngineBlockApiClientBundle\Exception;
+namespace OpenConext\EngineBlockApiClient\Exception;
 
-class ResourceNotFoundException extends RuntimeException
+use RuntimeException as CoreRuntimeException;
+
+class RuntimeException extends CoreRuntimeException implements OpenConextEngineBlockApiClientException
 {
 }
